Implement OTP verification code page
