+++ package.json
@@ -17,6 +17,7 @@
         "react": "^18.1.0",
         "react-dom": "^18.1.0",
         "react-dropzone": "^14.2.1",
+        "react-hook-form": "^7.32.1",
         "react-icons": "^4.3.1",
         "react-redux": "^8.0.2",
         "react-router-dom": "^6.3.0",
